clean login code
