+++ src/App.tsx
@@ -1,6 +1,6 @@
 import { useContext } from "react";
 import { AuthContext } from "./context/Context";
-import Login from "./pages/Login/Login";
+import Login from "./pages/Login";
 import "./index.css";
 import { Layout } from "./components/layout";
 import { ThemeProvider } from "./context/ThemeProvider";
