+++ src/pages/ScoringRules/RulesTable/RulesTable.tsx
@@ -13,11 +13,19 @@ import type { ViewRulesFormValues } from "../RulesFilter/useRulesFilter";
 import { DynamicTable } from "../../../components/DynamicTable";
 import PopupLayout from "../../../components/Popup/LayoutPopup";
 import RulesPopupJsx from "../RulesForm/RulesPopupJsx";
+import FullScreenSpinner from "../../../components/FullScreenSpinner";
 
 const ViewIcon = React.lazy(() => import("../../../assets/svg/View.svg?react"));
 const EditIcon = React.lazy(() => import("../../../assets/svg/Edit.svg?react"));
 const DeleteIcon = React.lazy(
   () => import("../../../assets/svg/Delete.svg?react")
+);
+const PlusIcon = React.lazy(() => import("../../../assets/svg/plus.svg?react"));
+const LockIcon = React.lazy(
+  () => import("../../../assets/svg/LockIcon.svg?react")
+);
+const BackgroundCircle = React.lazy(
+  () => import("../../../assets/svg/BackgroundCircle.svg?react")
 );
 
 type Rule = {
@@ -165,7 +173,7 @@ const RuleMenu = ({
 export const RulesTable = () => {
   const {
     data = [],
-    isLoading,
+    loadingState,
     error,
     totalCount,
     currentPage,
@@ -250,51 +258,129 @@ export const RulesTable = () => {
     setCurrentPage(1);
   };
 
+  const isFilterActive = useMemo(
+    () => Object.keys(filters ?? {}).length > 0 || searchText.trim() !== "",
+    [filters, searchText]
+  );
+
+  if (loadingState === "loading") {
+    return <FullScreenSpinner />;
+  }
+
   return (
-    <DynamicTable<Rule>
-      data={data.map((item) => ({
-        id: item.id,
-        name: item.name ?? "",
-        description: item.description ?? "",
-        status: item.status as "ENABLED" | "DISABLED",
-        riskLevel: item.riskLevel as "Low" | "Medium" | "High",
-      }))}
-      columns={columns}
-      filterComponent={
-        <RulesFilterForm
-          isOpen={isFilterOpen}
-          closeDrawer={closeFilterModal}
-          filterData={filters}
-          handleSearchSubmit={handleSearchSubmit}
+    <div className="p-6 bg-white shadow-sm dark:bg-[#121418] dark:border-gray-800">
+      <div className="mb-6">
+        <div className="flex items-center justify-between pt-5">
+          <h2 className="text-lg font-semibold text-gray-900 dark:text-white">
+            Scoring Rules{" "}
+            <span className="ml-2 text-sm text-blue-600 bg-blue-100 px-2 py-0.5 rounded-full">
+              {totalCount} Rule
+              {totalCount !== 1 && "s"}
+            </span>
+          </h2>
+
+          {totalCount !== 0 && (
+            <button
+              onClick={handleAddNewRule}
+              className="bg-blue-700 hover:bg-blue-800 text-white px-4 py-2 rounded-[8px] text-sm font-medium w-[155px] h-10 flex items-center justify-center gap-2"
+            >
+              <PlusIcon className="w-[20px] h-[20px]" />
+              Add New Rule
+            </button>
+          )}
+        </div>
+
+        <p className="text-sm text-gray-500 mt-1">
+          Keep track of customers and their security levels.
+        </p>
+      </div>
+
+      {totalCount === 0 && !isFilterActive ? (
+        <div className="w-full h-[75vh] flex flex-col items-center justify-center rounded-md border">
+          {/* Wrapper for icon + background */}
+          <div className="relative flex items-center justify-center mb-6 w-[80px] h-[80px]">
+            {/* Background Circle positioned behind */}
+            <div className="absolute z-0 w-[80px] h-[80px] flex items-center justify-center">
+              <BackgroundCircle
+                className="
+            absolute
+            left-1/2 top-[28%]
+            -translate-x-1/2 -translate-y-1/2
+            w-[400px] sm:w-[400px] md:w-[400px] lg:w-[400px]
+            h-[400px]
+            pointer-events-none select-none
+            z-0
+          "
+              />
+            </div>
+
+            {/* Lock Icon in styled border */}
+            <div className="relative z-10 flex items-center justify-center bg-white border border-[#D5D7DA] rounded-[16px] gap-[8px] p-[4px]">
+              <div className="flex items-center justify-center bg-white border border-black/10 rounded-[12px] sm:w-[52px] sm:h-[52px] p-[12px] shadow-[0px_1px_2px_0px_#0000001A,0px_3px_3px_0px_#00000017]">
+                <LockIcon className="sm:w-[28px] sm:h-[28px]" />
+              </div>
+            </div>
+          </div>
+
+          {/* Title & Description */}
+          <h3 className="text-lg font-medium text-gray-900 mb-1 mt-[48px]">
+            Start adding new rules
+          </h3>
+          <p className="text-sm text-gray-500 mb-6">
+            You don’t have any rule yet. Start securing by adding new rules now.
+          </p>
+
+          {/* Add Button */}
+          <button
+            onClick={handleAddNewRule}
+            className="bg-blue-700 hover:bg-blue-800 text-white px-4 py-2 rounded-[8px] text-sm font-medium w-[352px] h-10 flex items-center justify-center gap-2"
+          >
+            <PlusIcon className="w-[20px] h-[20px]" />
+            Add New Rule
+          </button>
+        </div>
+      ) : (
+        <DynamicTable<Rule>
+          data={data.map((item) => ({
+            id: item.id,
+            name: item.name ?? "",
+            description: item.description ?? "",
+            status: item.status as "ENABLED" | "DISABLED",
+            riskLevel: item.riskLevel as "Low" | "Medium" | "High",
+          }))}
+          columns={columns}
+          filterComponent={
+            <RulesFilterForm
+              isOpen={isFilterOpen}
+              closeDrawer={closeFilterModal}
+              filterData={filters}
+              handleSearchSubmit={handleSearchSubmit}
+            />
+          }
+          totalCount={totalCount}
+          currentPage={currentPage}
+          itemsPerPage={itemsPerPage}
+          setCurrentPage={setCurrentPage}
+          onFilterStatus={onFilterStatus as (status: string) => void}
+          statusFilter={statusFilter}
+          onClearSearch={handleClearSearch}
+          onAddNewItem={handleAddNewRule}
+          title="Scoring Rules"
+          error={error}
+          searchText={searchText}
+          setSearchText={setSearchText}
+          openFilterModal={openFilterModal}
+          applyFilters={applyFilters}
+          searchPlaceholder="Search rules"
+          showStatusFilter={true}
+          statusFilterOptions={[
+            { key: "All", label: "View All" },
+            { key: "ENABLED", label: "Active" },
+            { key: "DISABLED", label: "Inactive" },
+          ]}
         />
-      }
-      totalCount={totalCount}
-      currentPage={currentPage}
-      itemsPerPage={itemsPerPage}
-      setCurrentPage={setCurrentPage}
-      onFilterStatus={onFilterStatus as (status: string) => void}
-      statusFilter={statusFilter}
-      onClearSearch={handleClearSearch}
-      onAddNewItem={handleAddNewRule}
-      isLoading={isLoading}
-      error={error}
-      title="Scoring Rules"
-      description="Keep track of customers and their security levels."
-      searchText={searchText}
-      setSearchText={setSearchText}
-      openFilterModal={openFilterModal}
-      applyFilters={applyFilters}
-      emptyStateMessage="Start adding new rules"
-      emptyStateDescription="You don’t have any rule yet. Start securing by adding new rules now."
-      searchPlaceholder="Search rules"
-      showStatusFilter={true}
-      filters={filters ?? {}}
-      statusFilterOptions={[
-        { key: "All", label: "View All" },
-        { key: "ENABLED", label: "Active" },
-        { key: "DISABLED", label: "Inactive" },
-      ]}
-    />
+      )}
+    </div>
   );
 };
