+++ src/routes/AppRouter.tsx
@@ -2,7 +2,7 @@ import { createBrowserRouter, Navigate } from "react-router-dom";
 import { AppRoutes } from "./AppRoutes";
 import App from "../App";
 import Home from "../pages/Home";
-import LoginForm from "../pages/Login/Login";
+import LoginForm from "../pages/Login";
 import { ConstantKeys } from "../constants/ConstantKeys.constants";
 import Support from "../pages/Support";
 import AboutUs from "../pages/AboutUs";
